Extract non-DEM baked geometry composer
